\name{squeezy-package}
\alias{squeezy-package}
\docType{package}
\title{
\packageTitle{squeezy}
}
\description{
\packageDescription{squeezy}
}
\details{
See \code{\link{squeezy}} for example code.
}
\author{
\packageAuthor{squeezy}

Maintainer: \packageMaintainer{squeezy}
}
\references{
Mirrelijn M. van Nee, Tim van de Brug, Mark A. van de Wiel, "Fast marginal likelihood estimation of penalties for group-adaptive elastic net.", 2021
}

