\name{squeezy-package}
\alias{squeezy-package}
\alias{squeezy}
\docType{package}
\title{
\packageTitle{squeezy}
}
\description{
\packageDescription{squeezy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{squeezy}
\packageIndices{squeezy}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{squeezy}

Maintainer: \packageMaintainer{squeezy}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
